Prevent unneeded rebuilds of the home page
Even though the home page was not visible, it still updated on every
action, becuase it was below the current page (in the navigator).

Converting the HomePageContentViewModel and the DaysViewModel to built
values provides generated == implementations, preventing rebuilds when
the state relevant to the home page didn't change (which doesn't usually
happen when you're on e.g. the calendar page)

I also converted the CalendarWeekViewModel to a built value. That did
not turn into a performance improvement, I decided to keep it because it
doesn't do any harm and might somehow be a _slight_ improvement.