log in locally during a pending network request
This change makes it so that the initial splash screen is shown for a
shorter period of time. Instead, the loading indicator on the home page
will be shown for a bit longer.

This has the effect that users will see the stored data quicker,
which is particularly useful if the login request takes a lot of time.